requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-extensions==4.1
django-sslserver==0.22
channels==4.2.0
channels-redis==4.2.0
whitenoise==6.9.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
python-dotenv==1.0.1
pillow==10.4.0
numpy==2.2.6
python-dateutil==2.9.0.post0
requests==2.32.3
google-genai==1.65.0
google-generativeai==0.8.4
pandas==2.2.3
scikit-learn==1.6.1
django-crontab==0.7.1